refactor(layouts): layoutsディレクトリを機能グループに再編成
- container.tsx を containers/ に移動
- footer.tsx を footers/ に移動（テストファイルも含む）
- page-header.tsx を page-headers/ に移動
- section.tsx を sections/ に移動
- 各機能グループに index.ts を作成
- カテゴリ直下の index.ts を削除
- 空の page-header/ ディレクトリを削除
- 14ファイルのインポートパスを新しい構造に更新
- docs/components-reference.md を新しい構造に合わせて更新
