+++ src/app/bookmarks/page.tsx
@@ -2,8 +2,8 @@ import { Metadata } from 'next'
 import { Suspense } from 'react'
 
 
-import { Container } from '@/components/layouts'
-import { PageHeader } from '@/components/layouts'
+import { Container } from '@/components/layouts/containers'
+import { PageHeader } from '@/components/layouts/page-headers'
 import { BookmarkFilter, BookmarkList, BookmarkListSkeleton } from '@/features/bookmark'
 import { makeWebsiteQuery } from '@/lib/hackersheet'
