+++ src/app/docs/page.tsx
@@ -1,8 +1,8 @@
 import { Metadata } from 'next'
 import { Suspense } from 'react'
 
-import { Container } from '@/components/layouts'
-import { PageHeader } from '@/components/layouts'
+import { Container } from '@/components/layouts/containers'
+import { PageHeader } from '@/components/layouts/page-headers'
 import { Skeleton } from '@/components/ui/skeleton'
 import {
   AllDocumentList,
