+++ src/app/page.tsx
@@ -2,7 +2,8 @@ import { Metadata } from 'next'
 import { Suspense } from 'react'
 
 import { getPicupSlugs } from '@/actions'
-import { Container, Section } from '@/components/layouts'
+import { Container } from '@/components/layouts/containers'
+import { Section } from '@/components/layouts/sections'
 import { Link } from '@/components/navigations/link'
 import { Profile } from '@/components/site'
 import { SITE_DESC, SITE_NAME, RECENT_DOCS_COUNT } from '@/constants'
