+++ src/app/tags/[tagName]/page.tsx
@@ -3,8 +3,8 @@ import { notFound } from 'next/navigation'
 
 
 import { MutedMessage } from '@/components/feedback/message'
-import { Container } from '@/components/layouts'
-import { PageHeader } from '@/components/layouts'
+import { Container } from '@/components/layouts/containers'
+import { PageHeader } from '@/components/layouts/page-headers'
 import { DocumentList } from '@/features/document'
 import { ColorCircle, SmallTag } from '@/features/tag'
 import { client } from '@/lib/hackersheet'
