+++ src/app/tools/page.tsx
@@ -1,7 +1,8 @@
 import { Metadata } from 'next'
 import Link from 'next/link'
 
-import { Container, PageHeader } from '@/components/layouts'
+import { Container } from '@/components/layouts/containers'
+import { PageHeader } from '@/components/layouts/page-headers'
 
 const title = 'Tools'
 const description = '便利ツールをまとめた一覧ページです。'
