+++ src/components/layouts/footers/footer.tsx
