--- src/components/layouts/index.ts
@@ -1,4 +1,0 @@
-export { default as Container } from './container'
-export { default as Section } from './section'
-export { default as PageHeader } from './page-header'
-export { default as Footer } from './footer'
