+++ src/components/layouts/page-headers/index.ts
@@ -1,0 +1,1 @@
+export { default as PageHeader } from './page-header'
