+++ src/components/layouts/sections/index.ts
@@ -1,0 +1,1 @@
+export { default as Section } from './section'
